+SoKV3 ~SoKHTML
Added Saviour of Katar Version 3, this version is almost final and only
minor changes will be made due to the results of testing. Edited the
html file so the
title reads "Saviour of Katar" instead of "STARTPROGRAM"

+SoKversion3 +LinktoSOKv3

Added Saviour of Katar version 3 to the site.
This version is complete
and only minor alterations will be made from
this version. Added a link
to Saviour of Katar version 3 inside of my
portfolio page.